+++ src/services/userService.php
@@ -5,6 +5,8 @@ namespace services;
 require_once ROOT_DIR . 'models/userModel.php';
 require_once ROOT_DIR . 'repositories/userRepository.php';
 
+use exceptions\BadRequestException;
+use exceptions\UnauthenticatedException;
 use models\UserModel;
 use repositories\UserRepository;
 
@@ -28,54 +30,41 @@ class UserService
         $this->userRepo = userRepository::getInstance();
     }
 
-    function login(string $username, string $password): array
+    function login(string $username, string $password): UserModel
     {
         $user = $this->userRepo->getUserByUsername($username);
 
         if ($user == null) {
-            return [null, "User not found"];
+            throw new UnauthenticatedException("Username not found");
         }
 
         if (!password_verify($password, $user->user_password)) {
-            return [null, "Username and password does not match"];
+            throw new UnauthenticatedException("Username and password do not match");
         }
 
-        return [$user, "Login successful"];
+        return $user;
     }
 
-    function register(string $username, string $password, bool $isAdmin = false): array
+    function register(string $username, string $password, bool $isAdmin = false): void
     {
-        $user = $this->userRepo->getUserByUsername($username);
-
-        if ($user !== null) {
-            return [400, "Username already used"];
+        if ($this->userRepo->getUserByUsername($username)) {
+            throw new BadRequestException("Username already used");
         }
 
         $hashedPassword = password_hash($password, PASSWORD_DEFAULT);
-        $user = $this->userRepo->createUser($username, $hashedPassword, $isAdmin);
-
-        if ($user) {
-            return [200, "Login successful"];
-        } else {
-            return [500, "Failed to add user"];
-        }
+        $this->userRepo->createUser($username, $hashedPassword, $isAdmin);
     }
 
-    function updateUser(string $userId, string $username, string $password, ?bool $isAdmin = null): ?UserModel
+    function updateUser(string $userId, string $username, string $password, ?bool $isAdmin = null): UserModel
     {
-        $user = $this->userRepo->getUserByUsername($username);
-
-        // TODO: notify duplicate
-        if ($user !== null) {
-            // return [400, "Username not found"];
-            return null;
+        if ($this->userRepo->getUserByUsername($username)) {
+            throw new BadRequestException("Username already used");
         }
 
         if ($password != "")
             $password = password_hash($password, PASSWORD_DEFAULT);
-        $user = $this->userRepo->updateUser($userId, $username, $password, $isAdmin);
 
-        return $user;
+        return $this->userRepo->updateUser($userId, $username, $password, $isAdmin);
     }
 
     function getByUserId(string $musicId): ?UserModel
@@ -88,20 +77,17 @@ class UserService
         return $this->userRepo->getAllUsers(null)[0];
     }
 
-    function getAllUsersAdmin(int $page): array
+    function getAllUsersPaged(int $page): array
     {
         return $this->userRepo->getAllUsers($page);
     }
 
-    function deleteUser(int $user_id): bool
+    function deleteUser(int $user_id): void
     {
-        return $this->userRepo->deleteUser($user_id);
+        if (!$this->userRepo->getUserById($user_id)) {
+            throw new BadRequestException("Username already used");
+        }
+        
+        $this->userRepo->deleteUser($user_id);
     }
-
-    // function createMusic(int $user_id, string $title, string $genre, array $musicFile, ?array $coverFile): ?MusicModel
-    // {
-    //     $music = $this->musicRepo->createMusic($title, $user_id, $genre, $musicFile, $coverFile);
-
-    //     return $music;
-    // }
 }
